Initial commit
Established Connection to postgres pgadmin database via connection string in app settings after installing the necessary nuget packages, being entity frameworks and tools then updating the database through the package manager console to create the DBs container. proceeding to establish the inital model for the contact class that consisted of the necessary properties that the contact is contrived of, adding in also a migration so the DB knows that  i want to create the next iteration of the Db updating DB to reflect the new contacts table that was created as a new property in the application database class of the type contact class that possesses the properties that make up the columns that it is made up of and shown on pgadmin.