Registered a Service
Registered an image Service by use of an Interface which is a mechanism that allows loose coupling between something that needs a behavior and the implementation of that behavior. It pass around a contract or an interface instead of a concrete class. in turn created a class to implement the image interface that was structured.
The purpose of interface is to define set of functionality that a class will need to implement if it needs to implement this here interface.