Display image and enhanced the index
Display image and enhanced the index. by use of JS data tables.net that was linked to the index to display. add functionality for building searching sorting and paging
look to see whats being pushed into the view from within the Contacts Controller via the Index action, injected a service into the View , the controller action is pushing the data of the image and the View will use an injected instance of the image service to turn the data into a viewable image. made edits into the IndexView, replaced the html displayfor into an img tag that takes in the decoded data that was stored in the database, being the ImageData and type and converting it to display the image and not the data. added a class to the img tag of fluid, to fit the image in to the size of the table cell.