Added data utility class
Added data utility class to the data folder holding two methods called GetConnectionSring and BuildConnectionString to be responsible for choosing which connection string to use to connect to the database. if running locally it will use local connection string, if on heroku, it will chop and form the database url environment variable into a viable ConnectionString.